+++ spinetoolbox/ui/tree_view_form.py
@@ -413,6 +413,6 @@ class Ui_MainWindow(object):
         self.actionRemove_selection.setShortcut(QtWidgets.QApplication.translate("MainWindow", "Ctrl+Del", None, -1))
         self.actionRestore_Dock_Widgets.setText(QtWidgets.QApplication.translate("MainWindow", "Restore Dock Widgets", None, -1))
 
-from spinetoolbox.widgets.custom_qtreeview import ObjectTreeView, CopyTreeView
 from spinetoolbox.widgets.custom_qtableview import AutoFilterCopyPasteTableView
+from spinetoolbox.widgets.custom_qtreeview import ObjectTreeView, CopyTreeView
 from spinetoolbox import resources_icons_rc