+++ CHANGELOG.md
@@ -15,13 +15,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - It is now possible to copy, paste and duplicate project items on the Design View.
 - New Setting: *Sticky selection in Graph View*. Enables users to select if they want to use multi-selection or 
   single selection in the Graph view Object tree when selecting items with the **left-mouse button**.
+- Projects can be saved to any directory
+- Project name can be changed in Settings
 
 ### Changed
 - spinetoolbox is now a Python package.
 - spinetoolbox.py has been moved to repository root. To start the app, use command 'python -m spinetoolbox'.
 - Tool templates are now called Tool specifications
+- File->Open Project opens a file dialog, where you can open projects by selecting an old 
+  <project_name>.proj file or a Spine Toolbox Project directory. Valid Spine Toolbox projects are 
+  decorated with the Spine logo.
+- When opening an old style project (.proj file), the project must be upgraded to a new style project 
+  by selecting a new directory for the project.
+- Project information is not saved to a <project_name>.proj file anymore. This information is now located 
+  in file <project_dir>/.spinetoolbox/project.json. Every Spine Toolbox project has this file.
+- Work directory is now a global setting instead of a project setting
 
 ### Deprecated
+- Saving project information to .proj files
 
 ### Removed